+++ src/backend/joanie/core/flows/order.py
@@ -5,6 +5,7 @@ from contextlib import suppress
 from django.apps import apps
 from django.utils import timezone
 
+from sentry_sdk import capture_exception
 from viewflow import fsm
 
 from joanie.core import enums
@@ -258,7 +259,12 @@ class OrderFlow:
             and target
             in [enums.ORDER_STATE_PENDING_PAYMENT, enums.ORDER_STATE_COMPLETED]
         ):
-            self.instance.enroll_user_to_course_run()
+            try:
+                # ruff : noqa : BLE001
+                # pylint: disable=broad-exception-caught
+                self.instance.enroll_user_to_course_run()
+            except Exception as error:
+                capture_exception(error)
 
         if target == enums.ORDER_STATE_CANCELED:
             self.instance.unenroll_user_from_course_runs()