+++ src/backend/joanie/signature/backends/base.py
@@ -7,6 +7,8 @@ from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.utils import timezone as django_timezone
 
+from sentry_sdk import capture_exception
+
 from joanie.core import models
 
 logger = getLogger(__name__)